billing-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.7
redis>=5.0.1
kafka-python>=2.0.2
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
alembic>=1.13.0
# OAuth2/OIDC
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
# Observabilidade
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
# opentelemetry-exporter-prometheus removido - não disponível no PyPI
# Usando prometheus-client diretamente para métricas
# Logging estruturado
python-json-logger>=2.0.7

